requirements.txt
python-docx==0.8.11
PyMuPDF==1.23.8
Pillow==10.2.0
opencv-python==4.9.0.80
transformers==4.37.2
torch==2.2.0
nltk==3.8.1
spacy==3.7.2
numpy==1.24.3
tensorflow==2.15.0
streamlit==1.31.1
chardet==5.2.0 